+++ src/libs/routes.js
@@ -2,7 +2,7 @@ import _ from 'lodash/fp'
 import { compile, pathToRegexp } from 'path-to-regexp'
 import { routeHandlersStore } from 'src/libs/state'
 import * as AzurePreview from 'src/pages/AzurePreview'
-import * as Projects from 'src/pages/billing/Billing'
+import * as Projects from 'src/pages/billing/List/List'
 import * as Environments from 'src/pages/Environments'
 import * as FeaturePreviews from 'src/pages/FeaturePreviews'
 import * as Group from 'src/pages/groups/Group'
