+++ src/pages/billing/List/GCPNewBillingProjectModal.ts
@@ -9,7 +9,7 @@ import colors from 'src/libs/colors'
 import { reportErrorAndRethrow } from 'src/libs/error'
 import { getUser } from 'src/libs/state'
 import * as Utils from 'src/libs/utils'
-import { billingProjectNameValidator } from 'src/pages/billing/Billing'
+import { billingProjectNameValidator } from 'src/pages/billing/billing-utils'
 import CreateGCPBillingProject from 'src/pages/billing/CreateGCPBillingProject'
 import { GoogleBillingAccount } from 'src/pages/billing/models/GoogleBillingAccount'
 import { validate } from 'validate.js'
