+++ src/pages/billing/Project.js
@@ -25,7 +25,7 @@ import * as StateHistory from 'src/libs/state-history'
 import * as Style from 'src/libs/style'
 import { topBarHeight } from 'src/libs/style'
 import * as Utils from 'src/libs/utils'
-import { billingRoles } from 'src/pages/billing/Billing'
+import { billingRoles } from 'src/pages/billing/billing-utils'
 import { ExternalLink } from 'src/pages/billing/NewBillingProjectWizard/StepWizard/ExternalLink'
 import { SpendReport } from 'src/pages/billing/SpendReport/SpendReport'
 import { cloudProviders } from 'src/pages/workspaces/workspace/analysis/utils/runtime-utils'
